feat(output): now supports fixed sets of keys alias keys and similar
now supports a fixed structure and a fixed set of keys. Also it supports merging multiple input
keys. There are multiple ways to output a table. And you can filter the output.

BREAKING CHANGE: now supports a fixed structure and a fixed set of keys. Also it supports merging
multiple input keys. There are multiple ways to output a table. And you can filter the output.